Split retired profile installer tests
